studio/requirements.txt
langgraph
langgraph-prebuilt
langchain-core
langchain-community
langchain-openai
firecrawl-py
python-dotenv
typing-extensions

